Revert "fixed previous issues plus added the first chart"
This reverts commit ffc6ddeec9e19b98eab62c7bb354abdd3bdd4c18.
