+++ index.html
@@ -2,10 +2,8 @@
 <html>
   <head>
     <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Mashrou Siha</title>
     <link rel="stylesheet" href="./public/main.css">
-    <link href="https://fonts.googleapis.com/css?family=Roboto" rel="stylesheet">
   </head>
   <body>
     <div id="root"></div>
