+++ package.json
@@ -27,7 +27,6 @@
     "babel-preset-react": "^6.24.1",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
-    "react-router": "^4.2.0",
     "react-router-dom": "^4.2.2",
     "webpack": "^3.11.0"
   },
