+++ src/components/Landing_page.js
@@ -1,14 +1,14 @@
 import React from 'react';
-import ListButtons from './List_buttons';
+import Abutton from './Main_button';
 const LandingPage = () => {
-  return (<div className="top-level-container">
+  return (<div className="container">
     <div className="header">
       <img className="header-image" src="public/assets/logo.png" alt="Galille society logo"/>
     </div>
-    <div className="first-page-introduction-div">
+    <div className="div-introduction">
       <h2>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit.</h2>
     </div>
-    <ListButtons />
+    <Abutton title="Smoking" img="../../public/assets/smoke-512.ico"/>
   </div>);
 }
