+++ src/components/Main_button.js
@@ -1,29 +1,13 @@
-import React, {Component} from 'react';
-import { withRouter } from 'react-router-dom';
-import { Link } from 'react-router-dom';
-class MainButton extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      style: "main-buttons flip"
-    };
-  }
-  componenetDidMount() {
-    this.setState({style: "main-buttons"});
+import React, { Component } from 'react';
+
+const MainButton = (props) => {
+    return (
+      <div className="main-buttons" onClick={() => props.onButtonClick()}>
+        <img className="iconImg" src={props.img} alt={props.title + "icons"}/>
+        <span>{props.title}</span>
+        <img className="arrow" src="./public/assets/more-blue.png" alt=""/>
+      </div>
+    );
   }
 
-  render() {
-
-    return (<div className={this.state.style} onClick={this.props.onButtonClick} >
-      <img className="iconImg" src={this.props.img} alt={this.props.title + "icons"}/> {
-        this.props.title.split(" ").map((text, i) => {
-          return <span key={i}>{text}</span>
-        })
-      }
-      <img className="arrow" src="./public/assets/more-blue.png" alt=""/>
-    </div>);
-
-  }
-}
-//onClick={this.handleLink(this.props.obj)withRouter
-export default  MainButton;
+export default MainButton;
