+++ webpack.config.js
@@ -7,9 +7,6 @@ module.exports = {
     path: path.resolve(__dirname, "public"),
     filename: "bundle.js"
   },
-  devServer : {
-    historyApiFallback : true
-  },
   module: {
     loaders: [
       {
